backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.1.0
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tablib==3.4.0
uritemplate==4.1.1
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
flake8==5.0.4
gunicorn==20.0.4